Texture loads now behave properly with subdirs (todo slash on beginning)
Particle editor appears to be running and fully functional

+++ ParticleLib/Emitter.cs
@@ -95,7 +95,7 @@ namespace ParticleLib
 		}
 
 
-		internal Particle []Update(int msDelta, out int numParticles)
+		internal Particle []Update(float msDelta, out int numParticles)
 		{
 			numParticles	=0;
